+++ tests/xmindparser_tests.py
@@ -1,8 +1,8 @@
 import logging
 from json import dumps
-from os.path import join, dirname
+from os.path import join, dirname, exists
 
-from xmindparser import xmind_to_dict, config
+from xmindparser import *
 from xmindparser.xreader import logger, set_logger_level
 
 file_path = join(dirname(__file__), 'test.xmind')
@@ -18,3 +18,12 @@ def test_xmind_to_dict_debug():
 def test_xmind_to_dict_default():
     d = xmind_to_dict(file_path)
     logger.info(dumps(d))
+
+
+def test_xmind_to_json():
+    if exists('test.json'):
+        os.remove('test.json')
+
+    xmind_to_json(file_path)
+    assert os.path.exists('test.json')
+    os.remove('test.json')