+++ CHANGELOG.md
@@ -3,4 +3,5 @@ Change Log
 
 1.0.0
 
-The baby version.
+- Support parse xmind to dict data type with Python.
+- Support parse xmind to json file type.